Extract the Conversion handoff and refactor methods out of Send
