+++ ConnectorNavisworks/ConnectorNavisworks/Bindings/ConnectorNavisworksBindings.cs
@@ -1,21 +1,13 @@
 using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using System.Windows.Forms;
 using Autodesk.Navisworks.Api;
 using DesktopUI2;
 using DesktopUI2.Models;
-using DesktopUI2.Models.Settings;
-using DesktopUI2.ViewModels;
 using Speckle.ConnectorNavisworks.NavisworksOptions;
-using Speckle.ConnectorNavisworks.Other;
 using Speckle.Core.Kits;
-using Speckle.Core.Logging;
 using Speckle.Core.Models;
 using static Speckle.ConnectorNavisworks.Other.SpeckleNavisworksUtilities;
-using Application = Autodesk.Navisworks.Api.Application;
-using Cursor = System.Windows.Forms.Cursor;
 using MenuItem = DesktopUI2.Models.MenuItem;
 
 namespace Speckle.ConnectorNavisworks.Bindings;
@@ -104,69 +96,4 @@ public partial class ConnectorBindingsNavisworks : ConnectorBindings
     // TODO!
     =>
     throw new NotImplementedException();
-
-  public async Task RetryLastConversionSend()
-  {
-    if (s_activeDoc == null)
-    {
-      return;
-    }
-
-    if (CachedConvertedElements == null || s_cachedCommit == null)
-    {
-      throw new SpeckleException("Cant retry last conversion: no cached conversion or commit found.");
-    }
-
-    if (s_cachedCommit is Collection commitObject)
-    {
-      // _isRetrying = true;
-
-      var applicationProgress = Application.BeginProgress("Retrying that send to Speckle.");
-      _progressBar = new ProgressInvoker(applicationProgress);
-      _progressViewModel = new ProgressViewModel();
-
-      commitObject.elements = CachedConvertedElements;
-
-      var state = s_cachedState;
-
-      _progressBar.BeginSubOperation(0.7, "Retrying cached conversion.");
-      _progressBar.EndSubOperation();
-
-      var objectId = await SendConvertedObjectsToSpeckle(state, commitObject).ConfigureAwait(false);
-
-      if (_progressViewModel.Report.OperationErrors.Count != 0)
-      {
-        ConnectorHelpers.DefaultSendErrorHandler("", _progressViewModel.Report.OperationErrors.Last());
-      }
-
-      _progressViewModel.CancellationToken.ThrowIfCancellationRequested();
-
-      state.Settings.Add(new CheckBoxSetting { Slug = "retrying", IsChecked = true });
-
-      string commitId;
-      try
-      {
-        commitId = await CreateCommit(state, objectId).ConfigureAwait(false);
-      }
-      finally
-      {
-        _progressBar.EndSubOperation();
-        Application.EndProgress();
-        Cursor.Current = Cursors.Default;
-      }
-
-      state.Settings.RemoveAll(x => x.Slug == "retrying");
-
-      if (string.IsNullOrEmpty(commitId))
-      {
-        return;
-      }
-    }
-
-    // nullify the cached conversion and commit on success.
-    s_cachedCommit = null;
-
-    CachedConvertedElements = null;
-    // _isRetrying = false;
-  }
 }
